sqlbase: move functions to column_type_properties.go
Release note: None
